Add Vim-Flavors.
I realized this morning that I can use Vundle and Vim-Flavors
simultaneously (duh).  Since Vim-Flavors is a superior packaging
system—in my option—I will gradually move over to it.  It also works
hand-in-hand with vim-vspec and testing plugins is a great idea.

I have an older branch where I attempted to move over wholesale to
Vim-Flavors but ran into plugins that did not support Vim-Flavors,
mostly due to versioning/tags.  I expect this attempt to be more
successful, which is why I'm doing this in my main branch.